Refactoring fMRI_utils script to prepare for addition of scenarios joining
