+++ requirements.txt
@@ -3,3 +3,6 @@ torch==1.13.1
 torchaudio
 torchvision
 cudatoolkit==11.6
+bids
+nilearn
+nibabel
